requirements.txt
async-timeout==3.0.1
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.1
certifi==2020.12.5
click==7.1.2
gast==0.2.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
numpy==1.20.0rc1
oauthlib==3.0.1
opt-einsum==3.3.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
pypiwin32==223
pywin32==300
tensorflow==2.1.0
tensorflow-estimator==2.4.0rc0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
Werkzeug==0.16.1
wincertstore==0.2
